+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## Release 2.0.2 - January 18, 2021
+### Fixed
+* Button: modified variant `secondary`, fixed ActionBar
+* FileAttachment: modified prop `fileTypeLabel`
+* BubbleChat: fixed UI bug, removed console log, map style of `type=system` to `variant=primaryLight`, improve prop `options`
+* PaginationItem: removed console log
+* Toast: fixed warning
+### Added
+* Breadcrumb: added prop `noHref`
+* Icon: added icon `fileImport`, `fileExport`, `bubbles`
+### Updated
+* rc-slider, react-slick, react-split-pane, react-toastify to fixed version
+
 ## Release 2.0.1 - December 15, 2020
 ### Fixed
 * BubbleChat: pass `avatar` prop value to prop `name` of Avatar, instead of `src`